Split up the code into modules to make it readable
