+++ .env-example
@@ -1,0 +1,6 @@
+NODE_ENV=dev
+APP_ID="YOUR_MORALIS_APP_ID"
+API_URL="https://deep-index.moralis.io/api/v2/ipfs/uploadFolder"
+API_KEY="YOUR_X_API_KEY"
+SERVER_URL="YOUR_MORALIS_SERVER_URL"
+MASTER_KEY="YOUR_MORALIS_MASTER_KEY"
