+++ package.json
@@ -12,6 +12,7 @@
     "all": "^0.0.0",
     "axios": "^0.23.0",
     "canvas": "^2.8.0",
+    "dotenv": "^10.0.0",
     "moralis": "^0.0.104",
     "request": "^2.88.2"
   }
